refactor(transaction): give each action its own module
The action code was split by phase -- apply, footprint, proto -- so
understanding one action meant reading four files and adding one meant
touching four match statements.

Each action now owns a module holding its definition, `apply`,
`footprint`, wire conversions, and tests. `Action` dispatches to them.
The phase modules keep what is genuinely shared: `apply` holds the
`ApplyState` the actions program against, `footprint` the coordinate
space and the conflict comparison, `proto` the envelope and dispatch.

No behavior change. The one difference is that `AddField` and `AddBase`
now check for a duplicate local token before bumping the id counter,
matching `AddFragment`; a duplicate token aborts the whole apply either
way.
